+++ src/ThreadzAPI/ThreadzAPI.ts
@@ -9,6 +9,7 @@ import { Interact } from '../Interact/index.js';
 
 import type { Declarations, WorkerOptions } from '../declare/types.js';
 import type { ThreadzAPIConstructorOptions, ThreadzAPIEvents, MappedWorkers } from './types.js';
+import { BackgroundThreadzWorker } from '../BackgroundThreadzWorker/index.js';
 
 /**
  * Access declared workers and data about them via this API returned from the `declare()` function.
@@ -20,6 +21,7 @@ import type { ThreadzAPIConstructorOptions, ThreadzAPIEvents, MappedWorkers } fr
  * api.location;
  * api.declarations;
  * api.interactWith();
+ * api.createBackgroundWorker();
  */
 export class ThreadzAPI<T extends Declarations = Declarations> extends TypedEmitter<ThreadzAPIEvents> {
     /**
@@ -104,6 +106,10 @@ export class ThreadzAPI<T extends Declarations = Declarations> extends TypedEmit
         return Interact.with(this.workers[name]);
     }
 
+    createBackgroundWorker({ options = {} }: { options?: WorkerOptions } = {}): BackgroundThreadzWorker<ThreadzAPI<T>> {
+        return new BackgroundThreadzWorker<ThreadzAPI<T>>({ options, location: this.location });
+    }
+
     #queueWorker<A>({
         name,
         args,
@@ -116,7 +122,7 @@ export class ThreadzAPI<T extends Declarations = Declarations> extends TypedEmit
         priority: boolean;
     }): Promise<A> {
         return new Promise((resolve, reject) => {
-            const worker = new ThreadzWorker({ priority, options, workerData: { name, args, location: this.location } });
+            const worker = new ThreadzWorker({ priority, options, workerData: { name, args, location: this.location, type: 'REGULAR' } });
 
             ThreadzWorkerPool.enqueue(worker);