+++ src/worker/index.ts
@@ -1,3 +1,4 @@
+import fs from 'node:fs';
 import { parentPort, workerData } from 'worker_threads';
 
 import { SUCCESS_PAYLOAD, ERROR_PAYLOAD } from './consts.js';
@@ -5,23 +6,37 @@ import { ThreadzAPI } from '../ThreadzAPI/index.js';
 
 import type { WorkerData } from './types.js';
 import type { Declarations } from '../declare/types.js';
+import { BackgroundWorkerCallPayload, BackgroundWorkerCallResponse } from '../BackgroundThreadzWorker/types.js';
 
-const main = async () => {
+const getApi = async () => {
     try {
-        const { name, location, args } = workerData as WorkerData;
+        const { location } = workerData as WorkerData;
 
-        const api = (await import(location)).default as ThreadzAPI<Declarations>;
+        // if (!fs.existsSync(location)) {
+        //     throw new Error("It seems that the specified declarations file doesn't exist.");
+        // }
+
+        const imports = await import(location);
+
+        const api = Object.values(imports).find((item) => item instanceof ThreadzAPI) as ThreadzAPI<Declarations>;
 
         if (!api || !api?.declarations) {
-            throw new Error("Make sure you've made your declarations the default export of the file they're in.");
+            throw new Error("Make sure you've correctly exported your declarations in a separate file.");
         }
+
+        return api;
+    } catch (error) {
+        throw new Error("It seems that the specified declarations file doesn't exist.");
+    }
+};
+
+const regular = async () => {
+    try {
+        const { name, args } = workerData as WorkerData;
+        const api = await getApi();
 
         if (!api.declarations?.[name]) {
             throw new Error('There is no worker by this name in the specified declarations file.');
-        }
-
-        if (!(api instanceof ThreadzAPI)) {
-            throw new Error('The default export of your declarations file must be a ThreadzAPI instance.');
         }
 
         const result = await api?.declarations?.[name]?.worker(...args);
@@ -34,4 +49,24 @@ const main = async () => {
     }
 };
 
-main();
+const background = async () => {
+    const api = await getApi();
+
+    parentPort.on('message', async ({ name, id, args, terminate }: BackgroundWorkerCallPayload) => {
+        if (terminate) process.exit();
+
+        if (!api.declarations?.[name]) {
+            throw new Error('There is no worker by this name in the specified declarations file.');
+        }
+
+        try {
+            const payload = await api.declarations?.[name]?.worker(...args);
+            parentPort.postMessage({ name, id, payload } as BackgroundWorkerCallResponse);
+        } catch (error) {
+            parentPort.postMessage({ name, id, payload: null, error: (error as Error).message });
+        }
+    });
+};
+
+if ((workerData as WorkerData).type === 'REGULAR') regular();
+if ((workerData as WorkerData).type === 'BACKGROUND') background();