+++ src/workerTools/index.ts
@@ -29,12 +29,16 @@ const sendCommunication = <T extends AcceptableDataType>(data: T | SharedMemoryT
 /**
  * If you have passed a message port to the worker (using the Interact API), listen for messages on the port with this function.
  *
+ * @returns A function which will stop listening on the parent port when called.
+ *
  * @example
  * workerTools.onCommunication<string>((message) => console.log(`received: ${message}`));
  */
-function onCommunication<T extends AcceptableDataType>(callback: (data: T) => void): void;
-function onCommunication<T extends SharedMemoryTransferObject>(callback: (data: T) => void): void;
-function onCommunication<T extends AcceptableDataType = AcceptableDataType>(callback: (data: T | SharedMemoryTransferObject) => void) {
+function onCommunication<T extends AcceptableDataType>(callback: (data: T) => void): () => void;
+function onCommunication<T extends SharedMemoryTransferObject>(callback: (data: T) => void): () => void;
+function onCommunication<T extends AcceptableDataType = AcceptableDataType>(
+    callback: (data: T | SharedMemoryTransferObject) => void
+): () => void {
     const { port } = workerData as WorkerData;
 
     if (!port) {
@@ -46,6 +50,10 @@ function onCommunication<T extends AcceptableDataType = AcceptableDataType>(call
     }
 
     port.on('message', callback);
+
+    return () => {
+        port.off('message', callback);
+    };
 }
 
 /**
@@ -56,9 +64,9 @@ function onCommunication<T extends AcceptableDataType = AcceptableDataType>(call
  *
  * console.log(data);
  */
-async function waitForCommunication<T extends AcceptableDataType>(assertion: (data: T) => data is T): Promise<void>;
-async function waitForCommunication<T extends SharedMemoryTransferObject>(assertion: (data: T) => data is T): Promise<void>;
-async function waitForCommunication<T extends AcceptableDataType = AcceptableDataType>(assertion: (data: T) => data is T) {
+async function waitForCommunication<T extends AcceptableDataType>(assertion: (data: T) => boolean): Promise<void>;
+async function waitForCommunication<T extends SharedMemoryTransferObject>(assertion: (data: T) => boolean): Promise<void>;
+async function waitForCommunication<T extends AcceptableDataType = AcceptableDataType>(assertion: (data: T) => boolean) {
     const { port } = workerData as WorkerData;
 
     if (!port) {
@@ -70,8 +78,9 @@ async function waitForCommunication<T extends AcceptableDataType = AcceptableDat
     }
 
     return new Promise((resolve) => {
-        port.on('message', async (data) => {
+        port.on('message', function callback(data) {
             if (assertion(data)) resolve(data);
+            port.off('message', callback);
         });
     });
 }
@@ -84,12 +93,13 @@ async function waitForCommunication<T extends AcceptableDataType = AcceptableDat
  *
  * console.log(data);
  */
-async function waitForParentMessage<T extends AcceptableDataType>(assertion: (data: T) => data is T): Promise<void>;
-async function waitForParentMessage<T extends SharedMemoryTransferObject>(assertion: (data: T) => data is T): Promise<void>;
-async function waitForParentMessage<T extends AcceptableDataType = AcceptableDataType>(assertion: (data: T) => data is T) {
+async function waitForParentMessage<T extends AcceptableDataType>(assertion: (data: T) => boolean): Promise<void>;
+async function waitForParentMessage<T extends SharedMemoryTransferObject>(assertion: (data: T) => boolean): Promise<void>;
+async function waitForParentMessage<T extends AcceptableDataType = AcceptableDataType>(assertion: (data: T) => boolean) {
     return new Promise((resolve) => {
-        parentPort.on('message', async (data) => {
+        parentPort.on('message', function callback(data) {
             if (assertion(data)) resolve(data);
+            parentPort.off('message', callback);
         });
     });
 }
@@ -117,17 +127,25 @@ const sendMessageToParent = <T extends AcceptableDataType>(data: T | SharedMemor
  *
  * @param callback Function to run any time a message is received from the parent thread.
  *
+ * @returns A function which will stop listening on the parent port when called.
+ *
  * @example
  * workerTools.onParentMessage((data) => console.log(data))
  */
-function onParentMessage<T extends AcceptableDataType>(callback: (data: T) => void): void;
-function onParentMessage<T extends SharedMemoryTransferObject>(callback: (data: T) => void): void;
-function onParentMessage<T extends AcceptableDataType = AcceptableDataType>(callback: (data: T | SharedMemoryTransferObject) => void) {
+function onParentMessage<T extends AcceptableDataType>(callback: (data: T) => void): () => void;
+function onParentMessage<T extends SharedMemoryTransferObject>(callback: (data: T) => void): () => void;
+function onParentMessage<T extends AcceptableDataType = AcceptableDataType>(
+    callback: (data: T | SharedMemoryTransferObject) => void
+): () => void {
     if (isMainThread) {
         throw new MyError(ERROR_CONFIG('Attempting to use a workerTool on the main thread. Not allowed.'));
     }
 
     parentPort.on('message', callback);
+
+    return () => {
+        parentPort.off('message', callback);
+    };
 }
 
 /**