+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Ability to export `ThreadzAPI` not just as default.
+- `BackgroundThreadzWorker` class.
+- `createBackgroundWorker()` function to `ThreadzAPI`.
+
+### Changed
+
+- `worker.ts` file to support both background workers and regular workers.
+- `SharedMemory` API `set` methods to only update changed bytes.
+
 ## [2.2.1] - 2022/25/08
 
 ### Fixed